refactor: add itemtype, itemutil
