+++ src/types/Item.ts
@@ -1,0 +1,11 @@
+﻿export type ItemType = "FILE" | "NOTE" | "LINK";
+
+export interface Item {
+  id: string;
+  title: string;
+  description: string;
+  type: ItemType;
+  categories: string[];
+  isSelected: boolean;
+  isFavorite: boolean;
+}
